requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
fastapi==0.111.0
uvicorn==0.30.1